Refactor requisition mutation
Refactors requisition mutation to follow pattern established in
https://github.com/openmsupply/remote-server/pull/376.
