+++ src/server/service/graphql/schema/mutations/requisition/mod.rs
@@ -1,0 +1,3 @@
+mod insert;
+
+pub use insert::*;
